✨ Working XML support
