--- gdi-cases-server.tests/Models/DtoTests.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using DeepEqual.Syntax;
-using gdi_cases_server.Modules.Cases.Models;
-
-namespace gdi_cases_server.tests.Models;
-
-/// <summary>
-/// Test converting between JsonDto stuff (for JSON API) and XmlDTO (from XML API)
-/// </summary>
-[TestClass]
-public class DtoTests: ModelTestBase
-{
-    [TestMethod]
-    public void DumpXml()
-    {
-        var b = GenerateSampleBundle();
-        // Console.WriteLine(ToXml(XmlDto.FromJsonDto(b)));
-        Console.WriteLine(ToXml(b));
-    }
-    [TestMethod]
-    public void FromJsonDto()
-    {
-        // The XmlDto and JsonDto classes are used to convert between similar represenations
-        var b = GenerateSampleBundle();
-
-        XmlDto.FromJsonDto(b).ShouldDeepEqual(b);
-    }
-
-    [TestMethod]
-    public void FromXmlDto()
-    {
-        var b = GenerateSampleBundle();
-        JsonDto.FromXmlDto(XmlDto.FromJsonDto(b)).ShouldDeepEqual(b);
-    }
-}
