--- gdi-cases-server/Modules/Cases/Controllers/UploadCasesResult.cs
@@ -1,6 +1,0 @@
-﻿// For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-namespace gdi_cases_server.Modules.Cases.Controllers;
-
-public class UploadCasesResult
-{
-}
