+++ gdi-cases-server/gdi-cases-server.csproj
@@ -28,7 +28,6 @@
     <None Remove="AspNetCore.Authentication.ApiKey" />
     <None Remove="Converters\" />
     <None Remove="XmlSupport\" />
-    <None Remove="Modules\Cases\Models\Xml\" />
     <None Remove="Modules\Cases\Models\Json\" />
     <None Remove="Swashbuckle.AspNetCore.Annotations" />
     <None Remove="Microsoft.AspNetCore.Mvc.Formatters.Xml" />
@@ -42,7 +41,6 @@
     <Folder Include="Authentication\" />
     <Folder Include="Converters\" />
     <Folder Include="XmlSupport\" />
-    <Folder Include="Modules\Cases\Models\Xml\" />
     <Folder Include="Modules\Cases\Models\Json\" />
   </ItemGroup>
 </Project>
